+++ sdk/src/main/java/com/sportradar/livedata/sdk/di/LiveScoutInjectionModule.java
@@ -2,6 +2,9 @@ package com.sportradar.livedata.sdk.di;
 
 import com.google.inject.*;
 import com.google.inject.Module;
+import com.sportradar.livedata.sdk.proto.livescout.auth.AuthMessageProvider;
+import com.sportradar.livedata.sdk.proto.livescout.auth.TokenAuthMessageProvider;
+import com.sportradar.livedata.sdk.proto.livescout.auth.CredentialsAuthMessageProvider;
 import com.sportradar.livedata.sdk.common.classes.FileSdkLogger;
 import com.sportradar.livedata.sdk.common.classes.NullSdkLogger;
 import com.sportradar.livedata.sdk.common.exceptions.SdkException;
@@ -13,6 +16,7 @@ import com.sportradar.livedata.sdk.common.interfaces.SdkLogger;
 import com.sportradar.livedata.sdk.common.networking.ReconnectingGateway;
 import com.sportradar.livedata.sdk.common.networking.TcpGateway;
 import com.sportradar.livedata.sdk.common.rategate.*;
+import com.sportradar.livedata.sdk.common.settings.AuthSettings;
 import com.sportradar.livedata.sdk.common.settings.LimiterData;
 import com.sportradar.livedata.sdk.common.settings.LiveScoutSettings;
 import com.sportradar.livedata.sdk.common.timer.PeriodicTimer;
@@ -36,12 +40,13 @@ import com.sportradar.livedata.sdk.proto.dto.IncomingMessage;
 import com.sportradar.livedata.sdk.proto.dto.OutgoingMessage;
 import com.sportradar.livedata.sdk.proto.livescout.LiveScoutClientAliveProducer;
 import com.sportradar.livedata.sdk.proto.livescout.LiveScoutOutgoingMessageInspector;
-import com.sportradar.livedata.sdk.proto.livescout.LiveScoutStatusFactory;
+import com.sportradar.livedata.sdk.proto.livescout.LiveScoutOutgoingMessageFactory;
 import jakarta.xml.bind.JAXBContext;
 import org.apache.commons.net.DefaultSocketFactory;
 
 import jakarta.inject.Provider;
 import jakarta.inject.Singleton;
+
 import javax.net.ssl.SSLSocketFactory;
 import jakarta.xml.bind.JAXBException;
 import org.joda.time.Duration;
@@ -106,7 +111,7 @@ public class LiveScoutInjectionModule implements Module {
                 .to(LiveScoutEntityMapper.class)
                 .in(Singleton.class);
 
-        binder.bind(LiveScoutStatusFactory.class)
+        binder.bind(LiveScoutOutgoingMessageFactory.class)
                 .in(Singleton.class);
 
 
@@ -161,7 +166,7 @@ public class LiveScoutInjectionModule implements Module {
 
     @Provides
     @Singleton
-    protected LiveScoutUserRequestManagerImpl provideUserRequestManager(LiveScoutStatusFactory factory,
+    protected LiveScoutUserRequestManagerImpl provideUserRequestManager(LiveScoutOutgoingMessageFactory factory,
                                                                         TestManager testManager) {
         return new LiveScoutUserRequestManagerImpl(
                 USER_REQUEST_MANAGER_INDEX,
@@ -173,7 +178,7 @@ public class LiveScoutInjectionModule implements Module {
     @Provides
     @Singleton
     protected LiveScoutClientAliveProducer provideAliveProducer(
-            LiveScoutStatusFactory entityFactory,
+            LiveScoutOutgoingMessageFactory entityFactory,
             ScheduledExecutorService executor) {
         return new LiveScoutClientAliveProducer(entityFactory, executor, settings.getClientAliveMsgTimeout());
     }
@@ -238,9 +243,10 @@ public class LiveScoutInjectionModule implements Module {
     @Provides
     @Singleton
     protected Protocol<IncomingMessage, OutgoingMessage> provideProtocol(
-            LiveScoutStatusFactory statusFactory,
+            AuthMessageProvider authFactory,
             ScheduledExecutorService scheduledExecutorService,
-            Gateway gateway) throws JAXBException {
+            Gateway gateway
+    ) throws JAXBException {
 
         JAXBContext incomingContext = JAXBContext.newInstance(INCOMING_PACKAGE_NAME);
         JaxbBuilder incomingBuilder = new JaxbFactory(incomingContext);
@@ -262,9 +268,19 @@ public class LiveScoutInjectionModule implements Module {
                 messageWriter,
                 rateLimiter,
                 new LiveScoutOutgoingMessageInspector(),
-                statusFactory,
-                settings);
+                authFactory);
 
+    }
+
+    @Provides
+    @Singleton
+    protected AuthMessageProvider provideAuthMessageFactory(LiveScoutOutgoingMessageFactory factory) throws Exception {
+        AuthSettings authSettings = settings.getAuthSettings();
+        if (authSettings.isTokenAuth()) {
+            return new TokenAuthMessageProvider(factory, authSettings);
+        } else {
+            return new CredentialsAuthMessageProvider(factory, authSettings);
+        }
     }
 
     @Provides